fix: download artifacts from unmerged-lane-id when applicable (#6537)
When an artifacts is needed, Bit goes to the lane-remote or main-remote depends whether this version is on the lane.
In case a component is during merge, its aspects config/data consist of both: the current version and the merged version, as such, it's hard to know where to get the artifacts from. the current-version scope or the merged-version scope.
This PR attempts to go to the merged-scope first, and then to the current-version scope.