chore: update babel packages to latest (#6604)
It fixes the error we started getting on master today:
```
Error: [BABEL] /private/var/folders/dg/459ksxcn5cxdm0t9195t629h0000gn/T/bit/e2e/xykzv6on-local/foo.ts: You gave us a visitor for the node type TSSatisfiesExpression but it's not a valid type
```
The reason for this error is that some of the packages were updated to the newly released version 7.20.0, which had this new  `TSSatisfiesExpression` node type, and some of the packages were older and didn't recognize this type. 
This PR aligns them all to the latest.