perf: make history traversal faster by caching them in the fs (#6606)
Some commands, such as `bit status` traverse component history to get valuable data such as the distance between the local and the remote. This operation is pretty expensive as it requires to read each Version object from the filesystem to get its parent. 
This PR caches per component all the versions history-data it found. In some scenarios, when on a lane with lots of components and versions, it makes `bit status` 2.5 times faster.