fix(reset): change the head to be the common-before-diverge instead of the remote-head (#6692)
Currently, in case of diverge, during reset, the head (of the component or inside the lane) is set to the remote-head. Inside the .bitmap the version is set to be the same as the head.
This PR fixes it to set the head back to what it was before the local snaps as expected.