fix(export): avoid sending historical Version objects during export on lane (#6690)
currently, when exporting a lane with external components (components on a lane that belong to other scopes), there is a calculation of additional objects needed to be sent to the lane-scope in case it's missing from there.
This PR takes a different approach. It only sends the local-snapped objects. Other missing objects on the lane-scope, will be fetched directly by the lane-scope during the validation phase.