improvement: show missing packages when manually added with versions (#6702)
currently, when running `bit deps set` and the component dependency is not imported nor installed and is added with a specific version, it is added as package instead of a component.
It's happening because Bit doesn't have the package.json to know that it's a componnet.
This PR shows the component/package as missing until it's installed. Once installed, it's recognized correctly.