perf(import): avoid fetching flattened dependencies (#6705)
Currently, when importing a component, all its flattened (recursive) dependencies are fetched and saved into the local scope. 
The flattened dependencies were mostly needed for creating a graph dependencies. However, in many cases no need for a component-graph, it's enough to have a component-id graph. For this, it's possible to use the recently added "flattened-edges" prop in the Version object. It's providing essentially the entire dependencies graph of a component without the need to go to the remote and fetch them.
With this PR, when importing a component, and the component was tagged/snapped recently, it has the graph built-in and there is no need to fetch the deps. 
It also changes a few places where the component-graph was used to component-id graph.