fix: load extensions only once after merging from different sources (#6769)
Bit calculates the aspect based on different sources, such as, model, workspace.json, bitmap. At each source, the aspects are loaded. The reason was to get the envs registered so then we know whether we have already an env configured, and then ignore other envs from different sources.
This is now not needed, because all envs are registered via `teambit.envs/envs`, so it's possible to get the env from there and no need to load the aspect to check whether it's an env.
The load is done at the end, once all aspects from different sources are merged.