improvement: cache staged-snaps hashes to avoid fetching them from a remote (#7003)
Currently, to get this info whether or not a version is local, we have to run the expensive `getDivergeData` call.
Not only this is not good for performance reasons, but this is also throw errors in some cases when the version we search for is coming from another lane, not the currently checked out one. 
Also, in these cases, it's unable to determine whether it's local, because the `setDivergeData` only checks the currently checked out lane against the remote.