fix: avoid deleting the "version" prop from the package.json in the workspace (#6968)
In some scenarios, the `version` property is needed for the component to work properly. An example is when developing a VSCode extension. VSCode expects to have a valid package.json during the development.
To keep the `version` prop up to date, this PR also runs the `link` process (which creates the package.json) during watch and upon completion of tag/snap.