feat(pkg): allow custom-package-name without publishing to npm (#7027)
Currently, if the `packageJson.name` prop is set in the pkg aspect, the tag pipeline publishes the component to npm. There are cases, when the custom-name is needed without publishing, e.g. when a component is a VSCode extension. For these scenarios, a new config-prop is introduced `avoidPublishToNPM` to stop publishing.