fix(merge-from-scope), push artifacts to original scopes (#7119)
Currently, when merging lanes from a bare-scope, it doesn't import the
artifacts. As a result, when it pushes the components after the merge to
main or to another lane, the artifacts are missing from that scope.
It wasn't throwing an error during the export about missing artifacts
because it assumed they're all external and exist in the scope.
This PR imports the missing artifacts and also throw an error during the
export if they're missing.