fix(reset), fix VersionNotFound error when reseting a lane after merge (#7121)
This happens when due to the merge, many snaps arrived from the other
lane. By default, these `Version` objects are not fetched. When
resetting, all these snaps are "removed" locally and the head is changed
to how it was before the merge (more accurately, the recent export).
This PR loads these Version objects only when necessary. In most cases,
they're not needed.