fix(status): avoid showing components as pending-update when their remote-lane is empty (#7138)
currently, if the remote-lane doesn't exist, or it doesn't have this
specific component, it falls back to the either the forked-lane or to
the `component.head`. For bit-status or bit-checkout-head this is
incorrect. The comparison needs to be against the remote-lane, and if it
doesn't exist there, it should assume that the local is ahead.