fix(import): fetch versions that are shown in .bitmap but not in the lane object (#7167)
currently, when checked out to a lane, `bit import` fetches main
components from main and lane components according to the versions in
the lane.
In some scenarios, the .bitmap can be out-of-sync and a version in
.bitmap is different than a version in the lane object. Or maybe `bit
checkout` was running to a different version. In such cases, if the
local scope is deleted, bit was throwing `ComponentsPendingImport`
error.
This PR fixes the import process to fetch these versions as well.