improvement, save the flattenedEdges as a ref inside Version object (#7181)
currently, the `flattenedEdges` (graph of flattened dependencies) is
saved as a (potentially) huge json object in the `Version` model. It's
bad in terms of performance, because we parse the Version object very
often. It also makes it difficult to read when running `bit cat-object`
of the Version object.

This PR saves this JSON as a separate file (same as any other
source/artifact files) and keeps a Ref inside the Version object.
This also helps saving disc space because same content has the same
hash, as a result, multiple versions of the component when the
dependencies were not changed, will use the same file for the flattened
edges.