fix, enable adding a component on main when it was created on a lane (#7192)
Although the history tree is different.
Currently it shows an error "getHeadRegardlessOfLaneAsTagOrHash() failed
finding a head", which is not helpful.
This PR allow adding it and shows a warning suggests to remove the
component and merge it from the lane.

Also, in case of merging with --resolve-unrelated and current is main,
it throws another unhelpful error "localLane must be defined when
resolvedUnrelated". This is fixed to block running this flag when on
main. To resolve unrelated, the user needs to be on the problematic lane
and fix it from there. Main history always kept intact.

Co-authored-by: Gilad Shoham <shoham.gilad@gmail.com>