fix(merge): avoid writing failed snap versions in .bitmap file (#7185)
Normally, when tagging/snapping, we write the .bitmap to the filesystem
only once the process was completed successfully. This way, if the
build-process fails, it throws and exit the process without writing the
versions to the .bitmap.
In case of merge, the snap process is wrapped inside try/catch block, so
if it failed to run the pipeline, it still continues to the next step
and eventually writes the .bitmap at the end of the process with the
failed versions.
This PR fixes it by restoring the .bitmap content inside the `catch`
block.