fix, avoid overriding mutated Version objects (#7216)
In some cases, the `Version` object is mutated. For example when merging
with squash or when building on CI.
Currently, the object-writer is not aware of that and treat this object
as immutable. As a result, it's possible that locally a `Version` object
has changed and has the `squashed` prop and `parents` modified, but an
import process that fetched this Version as a dependency of another
component, overrode it.
This PR fixes it by adding a new `modified` property to the `Version`
object and it allows overriding only in case the incoming object is
newer.