fix(import), avoid throwing ComponentNotFound when .bitmap component was not found (#7223)
This is an out-of-sync case which apparently we didn't handle.
It happens when a component is exported. One user tagged the exported component locally, committed and pushed, another user clones the repo, gets the .bitmap with the versions according to the tagged one, which only exist on the first-user local but not on the remote.
Until now, in this case, `bit status` was showing the import-pending message, suggesting to import, but `bit import` was throwing `ComponentNotFound` error.
This PR fixes the import to not throw. Instead, in the output it shows the not-found components and suggests running `bit status` to sync the .bitmap back to the remote.