fix(status), avoid showing pending-merge when an out-of-date component was merged from another lane (#7285)
The use-case is when lane-a has a component that is out of date against
main. This lane is now merged into lane-b and a new snap is added to
this component.
Because lane-b is not a fork of lane-a, when comparing the local vs
remote for the diverge-data, we compare this component against main.
When comparing against main, the component is indeed diverged, however,
it should not be in "pending merge" section of bit-status. Pending merge
is only when a component is diverged against the remote scope. Not
against main.
This PR fixes it by not using the default `getDivergeData` and instead
comparing always to the remote-scope.