fix, fix aspects dependencies to get the version from .bitmap (#7295)
In some cases, when an aspect dependency is part of the workspace, the
component that uses it gets it with a version from the model instead of
the version from .bitmap. When running `bit show`, it shows the correct
version in the `dependencies` section, but the incorrect version in the
`aspect` and `devDependencies` section.
This PR sets the correct version from the .bitmap once the aspects are
calculated during the component-load.