fix(lane-merge), avoid showing workspace conflicts when it Semver satisfied (#7351)
## Proposed Changes

- Currently, `semver.satisfies` was checked only when the current-policy
and other-policy were a range, which never happened, because saving the
policy upon tag wasn't implemented yet. This PR removes this condition
and always check whether it satisfies, and if it is, it won't be shown
as a conflict.
- Fixes a bug where other-lane had the same version as workspace.jsonc,
but different version than current-component. It was showing the same
version from current/other as a conflict.