fix(out-of-sync): recreate a lane if local scope got deleted and is new (#7329)
Currently, in this case, the .bitmap still has the lane-id but the scope
doesn't have the lane object, so the user might be under impression that
the workspace is checked out to a lane. Snap command doesn't throw and
is snapped on main. Once the user is trying to export, an error is
thrown about missing lane object.
This PR re-create the lane in this case. since the .bitmap has the full
lane-id, the data to recreate the lane exists and it's possible to do.