fix, remove lane-refs from the cache once they are written (#7386)
Currently, when a ref file (.bit/refs/...) is loaded, it kept in memory
forever (until the process exits), and then when the refs are written,
it is re-written even if it wasn't changed. For long processes, such as
`bit start`, this can be an issue as it may write old values.
This PR fixes it by removing the refs from the cache after writing them
to the filesystem.