fix, support merging multiple ModelComponent from different remotes (#7401)
With the current mechanism of writing objects/components from different
remotes concurrently, it's possible that the same `ModelComponent`
arrived from different remotes (e.g. when fetching with dependencies).
Because it writes these components to the filesystem concurrently as
well, it doesn't "merge" them with each other only with the copy on the
filesystem. As a result, they can override each other and the incorrect
component could be written to the filesystem.

This PR introduces a new mechanism to write these component objects. It
keeps them all in one object until the "fetch" process is completed for
all remotes. Then, for each one of the received components, it loads the
copy from the fs, and merge them all with this copy and with each other
(unless we got a copy from the origin, which we can just ignore the
rest). Only once all the merge process is complete for all components,
it writes them to the filesystem, with no duplications and with the
correct merged components.