fix(status): fetch and save original lane when --lanes was used (#7425)
Previously, this fetch done during `bit status --lanes` was using
`importWithoutDeps` API, which was going to the remote without the
lane-id so it couldn't find the latest from the lane. Also, this API
doesn't merge the lane-object.
Instead, this PR uses `fetchLaneWithItsComponents`, which properly saves
the lane from the remote. Also, it improves this method to have less
roundtrips (2 instead of 3) and fetch less stuff (previously it was
fetching ids from main).