fix, avoid loading the version object for idOnLane (#7449)
`idOnLane` method is used to check whether a version exists on the lane
or on main. Until now, given the bitId, it was loading not only the
Component object but also the Version object, and if not exits it was
throwing ComponentNotFound.
It's definitely possible that the version in question doesn't exist
locally. This PR fixes it to only load the Component object.