fix(bin): update require to find hoisted dependencies
The previous `require` path in bin/bit.js depended on `@teambit/bit`
being in the `node_modules` dir within `@teambit/legacy`.

If the node dependency manager hoists `@teambit/bit` to be in the same
location as `@teambit/legacy`, then this `require` breaks.

I've only tested this locally, but it fixed an error I was getting in my
broken bit workspace installation when running any bit command