feat(lane-remove-comps): introduce a new sub-command "bit lane remove-comp" (#7558)
To clear the confusion about removing components from a lane.
Previously, removing components from a lane was done by running `bit
remove` and in most cases the `--delete` was needed. Now, this is the
default for the new `bit lane remove-comp` command.
In case the mark-remove is not needed, for example, when the lane is new
or the components are not part of the lane, it removes them from the
workspace without marking as deleted.