perf(get-snap-distance), huge performance improvement by switching R.difference with _.differenceBy (#7584)
For one component with 500 versions, it took 478ms with `R.difference`
and just 1ms with `Lodash.differenceBy`.
(this doesn't mean that Ramda is slower, probably Ramda has an
equivalent method. it was just probably the wrong way to compare two
huge arrays of objects without specifying the prop to compare).