fix(export), update VersionHistory if parents sent to the remote are missing (#7598)
This fixes an issue when `bit export` throws an error from the remote,
saying "missing data" of some parents snaps. It's happening when for
some reason the VersionHistory object is out of date and as a result,
`getDivergeData` is unable to connect the objects it got from the client
and the objects it has locally in the VersionHistory.
This PR re-populate the VersionHistory in this case with the missing
data.