default to not throw for NoCommonSnap (#7619)
In the following case the NoCommonSnap error blocks the user
unnecessarily.
lane-b forked from lane-a. 
lane-a continue snapping and then merged+squashed into main. 
On main the "squash" prop points to a newer version from lane-a, which
doesn't exist on lane-b.
On lane-b, getDivergeData compares its head to main, not to lane-a
because they're different scopes.
As a result, although it traverses the "squash", it's unable to connect
main to lane-b. The missing history exists on lane-a only.