warn when exporting/creating a forked lane to a different scope (#7628)
When the lane has a long history, it's possible that some of the history
is missing locally and when exporting this lane, it'll throw errors
about missing versions. Also, exporting such a lane may take a while as
it needs to push all that long history into the new scope.
This PR suggests the user to change the scope to the same scope of the
original lane if possible.