fix(remove), fix soft-removing a component after recovering it (#7724)
When a component is previously recovered, it has a prop "removed: false"
in the model, and then if this component is removed again, it has in
.bitmap "removed: true" of the same aspect `teambit.component/remove`.
Normally, .bitmap is strongest and it should show the component as
removed. However, due to merging of extensions from the legacy
ConsumerComponent, this is changed to `removed: false` as the legacy is
not aware of .bitmap config data.

This PR fixes it by removing the merge of the legacy data as it seems to
not be needed. It's possible that it was needed in the past and is not
needed anymore.