fix(switch), ensure components from other lanes are unavailable (#7706)
The trigger for this change is a bug when on a lane and importing a
component and snapping, the "onLanesOnly" prop in .bitmap doesn't get
added (coz technically it was available on main) , and then when
switching to another lane, this component is available unexpectedly. As
a result, the next snap could add this component as a dependency of
other components in a lane, which is invalid. A lane can't have
dependencies in other lanes.

Part of this fix is deprecating the `onLanesOnly` property. It's not
used anymore to determine which component is available on main. Instead,
we check whether a component-object (ModelComponent) has `head`. If it
does, we know that this component exist in main and we make it available
for you when switching to main even though you didn't use it on main
before. (it's kind of trade of, but in many cases, this is the desired
behavior regardless, and "bit checkout head" does it already by
default).

The reason for not removing `onLanesOnly` prop altogether, is to be
forward compatible. Otherwise, if users run recent bit version (which
strips this prop) and then using old version (which relies on this
prop), it will make all components available on main although they exist
only on lanes.

Another change in this PR is a big refactor of the switch code to reuse
the checkout API.