fix(env-get), remove wrong pipelines data, instead refer to using "bit build --list-tasks" (#7729)
This data of the pipelines coming from bit-env command is generated by
`BuilderService.getDescriptor()`.
One instance generates the data for the three pipelines: build, tag and
snap. However, the way how this class was design is to hold one pipeline
only.
It has a member `taskSlot`, which are the tasks registered to a
pipeline. It can be just one. build or snap or tag. Can’t be multiple.
What happened is that because only one instance was used, it showed the
same build registered tasks for all pipelines.

Instead of showing the task lists, it is now showing a message
suggesting the users to run `bit build --list-tasks`.