handle path for multiply components (#7821)
## Proposed Changes

**Context**:
We identified an issue with the `new-component-helper` when a path was
provided for multiple components. This led to a potential conflict,
causing failures for subsequent components due to existing paths.

**Enhancements**:
1. **Parameter Addition**: The `getNewComponentPath` function has been
extended with a third parameter to account for the number of components
being created. This ensures that the path returned is unique when
multiple components are being generated.

2. **Path Adjustments**: Depending on the number of components to create
and the existence of the dir, a path is created from the --path flag
alone or joined with the full name of the component.

3. **Synchronous Approach**: Made a conscious effort to keep the new
helper method synchronous to ensure that it doesn't inadvertently affect
or break functionality in other places where it might be used.

4. **Method Accessibility**: Considering the specific functionality and
intent of the `getNewComponentPath` method, I suggest that this method
should be private. It serves a very specific internal purpose, and
marking it as private would prevent unintended external access or
misuse.