introduce a new configuration to setup watcher to use polling (#7828)
On Mac, in some cases, when multiple watchers are running, the fsevents
doesn't respond to newly added files. (see
https://github.com/fsevents/fsevents/issues/394).
This configuration set `chokidar` to use-polling. The downside is that
it results in high CPU.

As a side note, I couldn't use the Chokidar `CHOKIDAR_USEPOLLING=true`
env variable, because while it does set the use-polling to true, but it
still keeps fsEvents to be `true`. This is not the same behavior when
passing `usePolling: true` to `chokidar.watch`, which sets the fsEvents
to `false`.