fix, delete merge-conflict file upon snap/tag even when it has workspace entries (#7845)
Currently, this file is deleted only once it's empty. However, it's
possible that this file still has the (relatively new) "WORKSPACE"
section, and as a result, it keeps the file, which makes it very
confusing later on to understand why some dependencies have specific
versions.
This PR checks whether all unmerged-components are now merged and if so
deletes the file.