update generator aspect (#7804)
### Pull Request: update-generator-aspect

**Summary:**

- **Introduced Generics:** Improved type flexibility by introducing
generics for both the `WorkspaceOptions` and `ComponentContext`. This
change allows users to define specific types for the `aspect` and
`template` fields without altering the core structure.
  
- **Conditionally Populate `.gitignore`:** Updated the starter template
logic to conditionally populate the `.gitignore` file based on the
`skipGit` flag in the `WorkspaceContext`.

- **Enhanced Component Templates:** Updated the component templates for
the starter and the component-generator to reflect the latest patterns,
ensuring consistency and adaptability.
  
- **Legacy Templates Update:** Legacy templates have been set to hidden,
and all legacy workspace templates have been removed from older core
environments.
  
**A Bit of Humor:**
It's been a unique challenge creating templates for templates! It's
almost like trying to explain a joke without ruining the
punchline...almost.

**Important Notes:**

- **Documentation Update:** Please hold off on merging this PR until the
documentation is ready. Expect it very soon!
  
**Motivation:**

The primary catalyst behind introducing generics was observing the
significant type duplication across the codebase. These duplications
were not only cumbersome to manage and update but also led to confusion.
By consolidating similar types through generics, we've reduced
redundancy and streamlined the type architecture.

Moreover, it was concerning that when developers use Bit, they
predominantly encounter legacy templates. This doesn't showcase the
capabilities of our new environments and might lead to a perception lag.
Hence, with this update, we aim to pivot developers towards the more
recent and efficient templates, providing them a taste of Bit's
cutting-edge capabilities right out of the gate.