fix(lane-merge), save versions in .bitmap as tags if exist when merging from main (#7891)
Currently, when merging from main and main is ahead, it saves the
components in .bitmap with the snap (hash) instead of the tags. As a
result, the subsequent snap of a dependent, saves the dependency with a
snap instead of a tag.
Once this happens, the manifest generated for this component has the
dependency version as `0.0.0-hash` instead of the tag and the
installation fails.
This PR fixes it by saving it with the tag to the .bitmap during the
merge.