fix(watch), use forked @teambit/chokidar and switch back to FsEvents (#7915)
This is a continuation of the watcher saga. See last update here:
https://github.com/teambit/bit/pull/7902
It may take some time for Chokidar owners to release a new version (the
last one was released Jan 22), so this repo was forked and a new version
was released under `@teambit/chokidar`.
With this new version it's possible to work with the FsEvents and avoid
the high CPU load.