fix(env-set), remove the previous env from .bitmap config root when it has no version (#7918)
This is relevant for custom tagged/exported local env. When components
in the workspace use this env, it gets saved in the .bitmap.config
without a version.
The issue was that when `bit env set` tries to remove this env, it gets
the full-id from the extension data including the version and search for
it in the .bitmap, which it couldn't find and therefore wasn't removing
it correctly.
This PR fixes it by always remove the instance without a version, and if
found an id with a version, remove it as well.