publish a new IPC event when objects are persist in the local scope (#7945)
A practical example when this is needed: bit-server is running in the
background (in vscode-ext), and the user is running `bit export` in the
CLI.
If this export triggered a .bitmap change (e.g. a component was new and
now exported), then bit-server knows to reload the consumer and clear
all caches.
If not, then, bit-server still have the old `ModelComponent` objects in
the scope cache, and it's not aware of the recent export.