support having multiple components with the same name but different scope-name in the same workspace (#7897)
This PR removes the restriction Bit always had of having components with
the same-name but different scope-name in one workspace.
For now, the case of multiple new components with the same-name is not
supported. They need to be exported.
Also, because it's experimental, it's disabled by default. To enable it,
use the bit-feature of `allow-same-name`.