support dynamic hash length (#8006)
Snaps use 40 characters hash. Until now, the CLI wasn't accepting a
shorter form, which in not convenience.
This PR checks whether a hash is shorter than 40 chars, and if so,
searches for an object starts with the given hash. If found more than
one, suggests the user to provide a longer hash.
In the output, currently, only "bit log --parents" shows short hashes,
but once this PR is merged, it's possible for other commands to show the
hash with various sizes.