replace legacy BitId with ComponentID (#8000)
The legacy `BitId` was used extensively not only in the legacy but also
in Harmony code.
The main constrain of using it was the "model"/"scope" layer, where
objects such as `ModelComponent` and `Version` were written historically
as `BitId`.
`ComponentID` is a wrapper around `BitId` which has all BitId
functionality and also makes the `scope` prop mandatory.
This what makes ComponentID superior. The ID is predictable. It always
has a scope-name. When parsing an ID from a string, we always know how
to parse because we know it has a scope-name.

Because this is a wrapper of BitId, it's easy to get the BitId instance
from ComponentID. However, to get a `ComponentID` instance from `BitId`
is very complex. In case BitId has no scope, we need to figure out what
is the scope (default-scope).

Another challenge we were constantly facing when changing or writing a
new `BitObject` is that when a component was never exported, the id was
written without scope-name (because it was an instance of `BitId`), and
then upon export, it was needed to go through each one of the those ids
and populate their scope-name.

One more limitation we had is that in one workspace you couldn't have
two new components with the same-name but different default-scope,
because their `BitId` instance is identical.

This PR simplifies/fixes all the above by using only `ComponentID` and
not BitId.