refactor, change Config class to return extensions from ConsumerComponent instance (#8147)
Currently, the `Config` class gets the `consumerComponent.extensions`
into the constructor. Problem is that in some cases, after the component
object is created, we replace the `consumerComponent.extensions` with
another ExtensionDataList instance. (happens in snap-from-scope
command). When this is happening, we get two unrelated instances of the
extensions, which leads to a very hard to debug extension issues.

This PR passes the instance of consumerComponent itself into the Config
constructor, and makes the `extensions` prop dynamic as a getter
function, so it's always up to date.