refactor, remove MissingComponent issue (#8210)
It's dead code. To find out whether a component is missing, we need to
identify that a package is really a component. To do this we go to the
package.json and search for the `componentID` property. If we found out
that the missing package is indeed a component, it means that we have
package.json so it's not missing.
This was relevant back then in the legacy when we conclude by the
package-name that it's a component. (by the `@bit`).