fix and simplify get-diverge-data (#8213)
The current algorithm is very complex. It has a recursive traversal of
the source, then on the target with lots of `if`s statements.

This new algorithm creates a graph and then subgraphs for the target and
source. Finds the diffs between the subgraph by simply comparing the
string arrays. Then, to find the common snap it traverses the graph by
BFS instead of DFS for efficiency.