refactor, split the legacy DependencyResolver to 3 parts: auto-detect, apply-overrides and version-resolver (#8217)
The end goal is to be able to call the apply-overrides from the scope.
It's needed for a future feature of creating new components from a
bare-scope.
Until now, all these three were interconnected in one huge file. It was
hard to figure out where a dependency is coming from and where its
version has been resolved.
This PR moves all version-resolution part that was in the
dependency-resolve to the existing `dependencies-versions-resolver.ts`
file (which didn't do much previously). It also extracts all the
auto-detection part which requires the workspace and node-modules to be
there into a new file `auto-detect`. Finally, it moves the code that
deals with different overrides (e.g. merge-config, deps from envs, deps
from bit-deps-set etc) to a new file `apply-overrides`. This file
doesn't require the consumer and can be run from a bare-scope.

As a bonus, a previous issue where `bit cc` was required after changing
`getDependencies` of a local env is not resolved. No need for the `bit
cc` anymore. This is because only the first heavy step of auto-detect is
cached in the filesystem. The resolution of the overrides is not.

One last thing, this entire dependency-resolver directory was extracted
from the legacy code and moved into Dependencies aspect, which result in
removing lots of ugly hooks from legacy to harmony.