perf, improve performance of scope.list() by parsing only the given types (#8225)
When running `bit cc`, the index.json is deleted, so any command
afterwards needs to rebuild this file. To build it, it needs to read all
local objects to check whether they're ModelComponent or Lane objects
and if so, add them to the index.
For large workspaces, this takes very long because it not only reads
every single file in the local scope but it also parses it.

This PR improves it by extracting the header after reading the file and
checking the type before parsing it. So then only for the given types
the parsing take place.

The improvement is huge. For a scope with 168,266 files, it was 107
seconds, and with this PR it's 24 seconds.