fix(create), restore LoadDeps hook after loading bit from the global scope (#8248)
`loadDeps` hook is registered when Dependencies aspect is instantiated.
The issue was that during `bit create`, we load another instance of bit
from the global scope. When this `loadBit` is happening, it creates a
new instance of Harmony and all aspects, including Dependencies aspect.
In this Harmony instance, there is no Workspace because it is loaded
from a global scope, hence the new registration of the `loadDeps` takes
place without workspace, which fails any component load afterwards
originated from the original Harmony instance.
These ugly Globals hooks will be removed once Legacy is gone. Until then
they keep bite us here and there.