fix(status,checkout-head), fix status to show deleted when merge-pending, fix checkout-head to consider deleted (#8256)
## Proposed Changes

- in case a component is both: diverge (merge-pending) and also deleted,
`bit status` shows it in both categories. (previously it was showing it
only as deleted).
- running `bit checkout head` on deleted component is not ignored
anymore. Instead, it is updated to the head while keeping the local
change of the removal. So the end result is that the files are the same
as the head and it's still locally deleted.
- allow running bit-reset on deleted components (otherwise, if the
component is diverged, there is no good way out of this state).