default to not re-fetch missing version history (#8270)
Currently, when import is running and a component is found to have an
incomplete version-history object, it refetches it.
This was originally done due to the new importer brings incomplete
version history objects. It should be fine now.
If for some reason it's needed, run
`BIT_FETCH_INCOMPLETE_VERSION_HISTORY=1 bit import`.