change "bit lane merge main" default to include non-lane components (#8279)
The scenario is when on a lane and some of the workspace components
belong to main and are not part of the lane.
In case these main components are out of date, how to update them?
In the past, `bit checkout main` was the way to go. This had been
changed to not include them, see
[here](https://github.com/teambit/bit/pull/6853) the reason.
Currently, `bit lane merge main` takes care of this components. However,
a change done 4 months ago, required a flag `--include-non-lane-comps`
to include them.
Turns out that this is confusing, and for the most cases, it is expected
to include them by default without any flag.
This PR does exactly this. In case these components are not wanted, a
new flag has introduced: `--exclude-non-lane-comps` to exclude them.