experimental - build the flattend-edges from previously saved graphs (#8281)
Currently, to build this graph on tag/snap, we import all flattened
dependencies.
With this approach, it's possible to skip the import and find the graph
in the previous flattenedEdges graphs.
This PR doesn't make the switch yet. It only adds the new method along
side the current method, and if an env variable is enabled, it compares
between the two and throws for discrepancies. The reason for the extra
cautions here is because it gets saved into the Version object and is
hard to fix if we get it wrong.