enable creating new components using snap-from-scope command (#8146)
Until now, the only way to create a new component was using the
workspace. The component files were physically on the filesystem and the
`Component` object was created using these files.
This PR implements a new way to create components - from the scope
without any workspace. It tackles a few challenges to make this work.
For example:
* lots of aspects data were calculated during some OnLoad slots which
were registered only when the workspace was available. This PR
introduces a new method `scope.registerOnCompAspectReCalc` to be able to
load the data of those aspects on scope.
* the composition data calculation was rely on the filesystem to read
the composition file and parse the AST. This PR decouple it and allow
the calculation to work on content in-memory.
* the entire dependencies calculation was tightly coupled to the
workspace. Most of the work to refactor it was done in
https://github.com/teambit/bit/pull/8217. In this PR the "auto-detect"
part was replaced by the dependencies received from the user and the
"overrides" part was calculated properly. This way, the generated
components could get the dependencies data from the env.